requirements.txt
django
djangorestframework
mysqlclient
django-cors-headers
django-simple-history